DEP-124: Added sample apps using Carthage/CocoaPods to install SDK
