--- Sample/Cartfile
@@ -1,3 +1,0 @@
-github "getyoti/yoti-doc-scan-ios" == 1.1.1
-github "getyoti/yoti-doc-capture-ios" == 1.5.5
-github "BlinkID/blinkid-ios" == 4.7.0
