--- Sample/Cartfile.resolved
@@ -1,3 +1,0 @@
-github "BlinkID/blinkid-ios" "v4.7.0"
-github "getyoti/yoti-doc-capture-ios" "v1.5.5"
-github "getyoti/yoti-doc-scan-ios" "v1.1.1"
