--- Sample/Sample/AppDelegate.swift
@@ -1,20 +1,0 @@
-//
-//  Yoti Doc Scan iOS SDK, Sample.
-//  Created by Barbara Rollet.
-//
-//  Copyright (c) Yoti Ltd 2019.
-//  All rights reserved.
-//  See LICENSE.md for license.
-//
-
-import UIKit
-
-@UIApplicationMain
-class AppDelegate: UIResponder, UIApplicationDelegate {
-    var window: UIWindow?
-
-    func application(_ application: UIApplication, didFinishLaunchingWithOptions
-        launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-        return true
-    }
-}
