+++ Samples/Carthage/Sample.xcodeproj/xcshareddata/IDETemplateMacros.plist
@@ -2,7 +2,9 @@
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
-	<key>IDEDidComputeMac32BitWarning</key>
-	<true/>
+	<key>FILEHEADER</key>
+	<string>
+//  Copyright © 2020 Yoti Ltd. All rights reserved.
+//</string>
 </dict>
 </plist>
