+++ Samples/Carthage/Sources/Sample/SupportingFiles/Assets.xcassets/AppIcon.appiconset/Contents.json
