+++ Samples/Carthage/Sources/Sample/SupportingFiles/Assets.xcassets/Contents.json
