+++ Samples/Carthage/Sources/Sample/SupportingFiles/Base.lproj/LaunchScreen.storyboard
