+++ Samples/Carthage/Sources/Sample/SupportingFiles/Base.lproj/Main.storyboard
@@ -1,11 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="14490.70" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="lDe-zc-9Ut">
-    <device id="retina4_7" orientation="portrait">
-        <adaptation id="fullscreen"/>
-    </device>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="15705" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="lDe-zc-9Ut">
+    <device id="retina4_7" orientation="portrait" appearance="light"/>
     <dependencies>
         <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="14490.49"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="15706"/>
         <capability name="Safe area layout guides" minToolsVersion="9.0"/>
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
@@ -35,7 +33,7 @@
                                     </userDefinedRuntimeAttribute>
                                 </userDefinedRuntimeAttributes>
                                 <connections>
-                                    <action selector="didSelectYotiDocScanButton" destination="lDe-zc-9Ut" eventType="touchUpInside" id="S0P-Gv-eZm"/>
+                                    <action selector="didSelectYotiSDKButton" destination="lDe-zc-9Ut" eventType="touchUpInside" id="ND9-Pw-VeB"/>
                                 </connections>
                             </button>
                         </subviews>
@@ -47,7 +45,7 @@
                         <viewLayoutGuide key="safeArea" id="WAz-IC-Dcv"/>
                     </view>
                     <connections>
-                        <outlet property="yotiDocScanButton" destination="2Ba-PN-qam" id="GO9-mu-kkr"/>
+                        <outlet property="yotiSDKButton" destination="2Ba-PN-qam" id="0re-4a-6M8"/>
                     </connections>
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="rqb-mt-4NM" userLabel="First Responder" sceneMemberID="firstResponder"/>
