+++ Samples/Carthage/Sources/Sample/SupportingFiles/Input.xcfilelist
@@ -1,0 +1,11 @@
+$(SRCROOT)/Carthage/Build/iOS/YotiFoundation.framework
+$(SRCROOT)/Carthage/Build/iOS/YotiNetwork.framework
+$(SRCROOT)/Carthage/Build/iOS/YotiCommon.framework
+$(SRCROOT)/Carthage/Build/iOS/YotiSDKCommon.framework
+$(SRCROOT)/Carthage/Build/iOS/YotiSDKDesign.framework
+$(SRCROOT)/Carthage/Build/iOS/YotiSDKCore.framework
+$(SRCROOT)/Carthage/Build/iOS/YotiSDKDocument.framework
+$(SRCROOT)/Carthage/Build/iOS/YotiSDKZoom.framework
+$(SRCROOT)/Carthage/Build/iOS/YotiDocumentCapture.framework
+$(SRCROOT)/Carthage/Build/iOS/ZoomAuthenticationHybrid.framework
+$(SRCROOT)/Carthage/Checkouts/blinkid-ios/Microblink.framework
