+++ Samples/CocoaPods/Sample.xcodeproj/xcshareddata/IDETemplateMacros.plist
@@ -4,12 +4,7 @@
 <dict>
 	<key>FILEHEADER</key>
 	<string>
-//  Yoti Doc Scan iOS SDK, Sample.
-//  Created by ___FULLUSERNAME___.
-//
-//  Copyright (c) Yoti Ltd ___YEAR___.
-//  All rights reserved.
-//  See LICENSE.md for license.
+//  Copyright © 2020 Yoti Ltd. All rights reserved.
 //</string>
 </dict>
 </plist>
