+++ Samples/CocoaPods/Sources/Sample/AppDelegate.swift
@@ -1,0 +1,16 @@
+//
+//  Copyright © 2020 Yoti Ltd. All rights reserved.
+//
+
+import UIKit
+
+@UIApplicationMain
+class AppDelegate: UIResponder, UIApplicationDelegate {
+    var window: UIWindow?
+
+    func application(
+        _ application: UIApplication,
+        didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+        true
+    }
+}
